requirements/common.txt
Django==1.9.2
django-bootstrap3==6.2.2
django-choices==1.4.2
django-input-mask==1.3.7.1
django-localflavor==1.2
